feat(verification): Lean oracle v0.4.0 with full-fidelity types and CRDT differential tests
Major verification infrastructure upgrade enabling formal verification of
Rust CRDT implementations against Lean specifications.

Lean Oracle Enhancements:
- Full-fidelity type system matching Rust structures (Journal, Fact, Namespace)
- 4-variant TimeStamp (OrderClock, Physical, Logical, Range)
- 4-variant FactContent (AttestedOp, Relational, Snapshot, RendezvousReceipt)
- Modular type files under verification/lean/Aura/Types/
- JSON serialization with explicit variant tags for unambiguous parsing

Differential Testing:
- Rust CRDT law tests: commutativity, associativity, idempotence
- Property-based tests over 100+ random journals
- OrderTime conversion between Lean and Rust for invariant comparison
- Both systems verified to satisfy same semilattice axioms

Type Safety Improvements:
- Comprehensive identifier type safety across crates
- Delta type cleanup and validation
- Guard chain type refinements

All 24 differential tests pass (4 Rust-native + 20 Lean oracle).

+++ crates/aura-sync/src/core/metrics.rs
@@ -976,8 +976,8 @@ mod tests {
                 collector_clone.record_sync_start(&format!("session_{i}"), now);
                 collector_clone.record_sync_completion(
                     &format!("session_{i}"),
-                    i,
-                    i * 100,
+                    i as u64,
+                    (i as u64) * 100,
                     now + 50,
                 );
             });
